Fixed spring-boot quickstarts to use jboss-jdk base image
Also merged springboot-webmvc and springboot-webmvc-ip into one, moved springboot-keycloak to sandbox.

Started to rename project names for consistencies sake.
