--- archetypes/java-jgroups-greeter-archetype/.gitignore
@@ -1,1 +1,0 @@
-src
