+++ archetypes/pom.xml
@@ -55,15 +55,12 @@
    	<module>karaf-camel-rest-sql-archetype</module>
    	<module>karaf-cxf-rest-archetype</module>
     <module>karaf4-camel-log-archetype</module>
-   	<module>spring-boot-activemq-archetype</module>
    	<module>spring-boot-camel-archetype</module>
-   	<module>spring-boot-keycloak-archetype</module>
    	<module>spring-boot-webmvc-archetype</module>
-   	<module>spring-boot-webmvc-ip-archetype</module>
    	<module>war-camel-servlet-archetype</module>
    	<module>war-cxf-cdi-servlet-archetype</module>
    	<module>war-rest-archetype</module>
-        <module>war-wildfly-archetype</module>
+    <module>war-wildfly-archetype</module>
   </modules>
 
   <build>
