--- archetypes/spring-boot-activemq-archetype/.gitignore
@@ -1,1 +1,0 @@
-src
