--- archetypes/spring-boot-keycloak-archetype/.gitignore
@@ -1,1 +1,0 @@
-src
