--- archetypes/spring-boot-keycloak-archetype/pom.xml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<!--
-
-     Copyright 2005-2015 Red Hat, Inc.
-
-     Red Hat licenses this file to you under the Apache License, version
-     2.0 (the "License"); you may not use this file except in compliance
-     with the License.  You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-     Unless required by applicable law or agreed to in writing, software
-     distributed under the License is distributed on an "AS IS" BASIS,
-     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
-     implied.  See the License for the specific language governing
-     permissions and limitations under the License.
-
---><project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>io.fabric8.archetypes</groupId>
-        <artifactId>archetypes</artifactId>
-        <version>2.2.36-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>spring-boot-keycloak-archetype</artifactId>
-    <name>Fabric8 :: Archetypes :: Spring-Boot :: Keycloak</name>
-    <packaging>maven-archetype</packaging>
-
-    <description>Creates a new Shows how to use Spring Boot with Keycloak</description>
-
-    <build>
-        <extensions>
-            <extension>
-                <groupId>org.apache.maven.archetype</groupId>
-                <artifactId>archetype-packaging</artifactId>
-                <version>2.2</version>
-            </extension>
-        </extensions>
-
-        <pluginManagement>
-            <plugins>
-                <plugin>
-                    <groupId>org.apache.maven.plugins</groupId>
-                    <artifactId>maven-archetype-plugin</artifactId>
-                    <version>2.2</version>
-                </plugin>
-            </plugins>
-        </pluginManagement>
-
-        <!-- include all resources but only filter archetype-metadata.xml -->
-        <resources>
-            <resource>
-                <directory>src/main/resources</directory>
-                <filtering>false</filtering>
-            </resource>
-            <resource>
-                <directory>src/main/resources-filtered</directory>
-                <filtering>true</filtering>
-                <includes>
-                    <include>**/archetype-metadata.xml</include>
-                </includes>
-            </resource>
-        </resources>
-    </build>
-
-</project>
