--- archetypes/spring-boot-webmvc-ip-archetype/.gitignore
@@ -1,1 +1,0 @@
-src
