+++ quickstarts/spring-boot/README.md
@@ -1,13 +1,11 @@
-Spring-Boot Quickstarts
-=======================
+# Spring-Boot Quickstarts
 
 Various quickstart projects using [Spring Boot](http://projects.spring.io/spring-boot).
 
 The following quickstarts is provided out of the box:
 
-* [activemq](activemq) - This example demonstrates how you can use Apache ActiveMQ with Spring Boot in a [Java Container](http://fabric8.io/gitbook/javaContainer.html)
-* [camel](camel) - This example demonstrates how you can use Apache Camel with Spring Boot in a [Java Container](http://fabric8.io/gitbook/javaContainer.html)
-* [keycloak](keycloak) - This example demonstrates how you can use Spring Boot and web mvc with a [Java Container](http://fabric8.io/gitbook/javaContainer.html), using Keycloak for authentication.
-* [webmvc](webmvc) - This example demonstrates how you can use Spring Boot and web mvc with a [Java Container](http://fabric8.io/gitbook/javaContainer.html)
-* [webmvc-ip](webmvc-ip) - This example demonstrates how you can use Spring Boot and web mvc with a [Java Container](http://fabric8.io/gitbook/javaContainer.html) and returning IP to demonstrate load balancing
+* [activemq](activemq) - This example demonstrates how you can use Apache ActiveMQ with Spring Boot 
+  based on a [Java base image](https://github.com/fabric8io/base-images#java-base-images)
+* [camel](camel) - This example demonstrates how you can use Apache Camel with Spring Boot 
+  based on a [Java base imager](https://github.com/fabric8io/base-images#java-base-images)
