--- quickstarts/spring-boot/webmvc-ip/README.md
@@ -1,10 +1,0 @@
-# Web MVC IP Address with Spring-Boot QuickStart
-
-This example assumes that you have a running openshift v3, and fabric8 v2 using docker.
-For more info see http://fabric8.io/v2/getStarted.html.
-
-This example demonstrates how you can use Spring Boot and web mvc with a [Java Container](http://fabric8.io/gitbook/javaContainer.html)
-
-The quickstart uses Spring Boot to configure a little Spring MVC application that offers a REST service which returns IP address to show service load-balancing
-
-When the application is running, you can use a web browser to access the REST service.
