--- quickstarts/spring-boot/webmvc-ip/src/main/java/io/fabric8/quickstarts/springbootwebmvcip/IPAddress.java
@@ -1,36 +1,0 @@
-/**
- *  Copyright 2005-2015 Red Hat, Inc.
- *
- *  Red Hat licenses this file to you under the Apache License, version
- *  2.0 (the "License"); you may not use this file except in compliance
- *  with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
- *  implied.  See the License for the specific language governing
- *  permissions and limitations under the License.
- */
-package io.fabric8.quickstarts.springbootwebmvcip;
-
-public class IPAddress {
-
-    private final long id;
-
-    private final String ipAddress;
-
-    public IPAddress(long id, String ipAddress) {
-        this.id = id;
-        this.ipAddress = ipAddress;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public String getIpAddress() {
-        return ipAddress;
-    }
-}
