--- quickstarts/spring-boot/webmvc-ip/src/main/resources/application.properties
@@ -1,1 +1,0 @@
-spring.main.sources = io.fabric8.quickstarts.springbootwebmvcip
