+++ quickstarts/spring-boot/webmvc/src/main/java/io/fabric8/quickstarts/springbootmvc/domain/Invoice.java
@@ -34,7 +34,6 @@ import static javax.persistence.GenerationType.AUTO;
 public class Invoice {
 
     // Members
-
     @Id
     @GeneratedValue(strategy = AUTO)
     private long id;
