+++ quickstarts/war/pom.xml
@@ -27,7 +27,7 @@
   </parent>
 
   <groupId>io.fabric8.quickstarts</groupId>
-  <artifactId>quickstart-war</artifactId>
+  <artifactId>quickstarts-war</artifactId>
   <version>2.2.36-SNAPSHOT</version>
   <packaging>pom</packaging>
   <name>Fabric8 :: Quickstarts :: War :: Parent</name>
