+++ sandbox/quickstarts/activemq/src/main/java/io/fabric8/quickstarts/activemq/App.java
