+++ sandbox/quickstarts/activemq/src/main/resources/application.properties
