+++ sandbox/quickstarts/keycloak/src/main/java/io/fabric8/quickstarts/springbootkeycloak/IPAddress.java
