+++ sandbox/quickstarts/keycloak/src/main/java/io/fabric8/quickstarts/springbootkeycloak/IPAddressController.java
