+++ sandbox/quickstarts/keycloak/src/main/resources/application.properties
