+++ sandbox/quickstarts/keycloak/src/test/java/io/fabric8/tests/integration/KubernetesIntegrationKT.java
